aws/gerar_sumula/requirements.txt
azure-cosmos
pandas
regex
python-dotenv